walk: add MaxDepth + functional options; misc linter cleanups
walk:
- New Options.MaxDepth (sign-encoded int: -1=Unbounded, 0=roots only,
  N>0=cap at depth N). Matches GNU `find -maxdepth 0`. Pruned at enqueue
  in doWalk/walkPath/doSymlink; at-cap dirs are emitted in place since
  the worker (which normally lstats+emits) never sees them.
- Functional-options layer: Option, NewOptions (seeds Unbounded), Apply,
  and WithFoo setters covering every Options field. Bool-taking setters
  (WithFollowSymlinks/WithOneFS/WithIgnoreDuplicateInode) so CLI-flag
  plumbing reads naturally.
- Existing walk.Options{} literals in walk/, cmp/, clone/ updated to set
  MaxDepth: walk.Unbounded. The Go zero value now means "roots only",
  not "unbounded" -- struct-literal callers must opt in explicitly.
- Three new depth tests: roots-only, bounded-at-2, unbounded-baseline.

Linter cleanups across the rest of the tree:
- cmp: rename IGN_* -> Ignore* and internal _D_* -> d* constants;
  err_L/err_R -> errL/errR; switch to walk.Func (formerly WalkFunc).
- info, xattr, safefile, copy_*, clone tests: assorted vet and
  staticcheck findings.
- Pin linter tooling: add linters.mod/linters.sum and gen.go;
  expand .golangci.yml.

+++ info_darbsd.go
@@ -26,7 +26,7 @@ func makeInfo(fi *Info, nm string, st *syscall.Stat_t, x Xattr) {
 	*fi = Info{
 		Ino:    st.Ino,
 		Siz:    st.Size,
-		Blocks: int64(st.Blocks),
+		Blocks: st.Blocks,
 		// st.Dev/st.Rdev are int32 on darwin and uint64 on
 		// freebsd. Real device IDs are non-negative; the
 		// uint32 step makes the zero-extension explicit on
